feat(packages/sui-ci): Move correctly imports to ESModules
