+++ packages/sui-ci/src/index.js
@@ -1,2 +1,2 @@
-export {default as updateCommitStatus} from './updateCommitStatus'
-export {default as release} from './release'
+export {default as updateCommitStatus} from './updateCommitStatus.js'
+export {default as release} from './release.js'
